feat: moved "manage-votecount" cmd to a general subcommand model
